Create interface.py
So this script currently does a few things right:
Prompting the runner to select the file to import. i.e. using nws10 from the Google Drive
Reading the file correctly and sending to a Dataframe

What it lacks is the ability to incorporate the fix_encoding function such that it updates each cell within the dataframe. That is the MAIN issue of this script.